refactor(component): leaf nodes as transforms
Convert internal terminate nodes to identity transform nodes to avoid
interfering on the run status. In the previous implementation,
internal Terminate nodes performed calls to update the run return status
and metadata, causing inconsistent results.

Renamed the output joiner to the original name of the component. This
allows us to remove the logic for dependency substitution, as the
"original dependency" will exist, albeit after being replaced.

Signed-off-by: Antonio Pedro Vieira <apsvieira95@gmail.com>